v2.5.6: AutoClicker checks, DiagnosticLogger, dual-marketplace update checker
- Add AutoClickerA (CV + 30 CPS cap), AutoClickerB (bucket-CV randomized bot detection), AutoClickerC (multi-attack-per-tick) checks
- Add DiagnosticLogger: opt-in per-player movement recorder to diagnose false flags vs real cheats; disabled by default
- Update checker now covers both Modrinth and bdevs.it with cross-platform version sync notifications
- Bump version 2.5.5 to 2.5.6

+++ common/src/main/java/ac/voidac/predictionengine/PlayerBaseTick.java
@@ -446,7 +446,7 @@ public final class PlayerBaseTick {
                     if (fluidHeight == 0)
                         continue;
 
-                    double d0 = (float) (y + 1) - fluidHeight;
+                    double d0 = (y + 1) - fluidHeight;
 
                     if (!player.isFlying && ceilY >= d0) {
                         hasPushed = true;
